Switch from ints to floats as the coordinate storage for glgui, for sub-pixel rendering in the gui.

+++ glgui/button.h
@@ -11,14 +11,14 @@ namespace glgui
 		friend class CSlidingPanel;
 
 	public:
-						CButton(int x, int y, int w, int h, const tstring& sText, bool bToggle = false, const tstring& sFont=_T("sans-serif"), size_t iSize=13);
+						CButton(float x, float y, float w, float h, const tstring& sText, bool bToggle = false, const tstring& sFont=_T("sans-serif"), size_t iSize=13);
 
 	public:
 		virtual void	Think();
 
 		virtual void	Paint() { CLabel::Paint(); };
-		virtual void	Paint(int x, int y, int w, int h);
-		virtual void	PaintButton(int x, int y, int w, int h);
+		virtual void	Paint(float x, float y, float w, float h);
+		virtual void	PaintButton(float x, float y, float w, float h);
 
 		virtual bool	MousePressed(int code, int mx, int my);
 		virtual bool	MouseReleased(int code, int mx, int my);
